Speed up the recently added "database is busy" SQLite3 test
It was passing but only after ~57 seconds because the busy_timeout setting in
the test definition was being applied to a different connection than was being
used by the implementation.

With this change we hook into the connection function and apply the shorter
busy_timeout after the normal setup work is done but before handing the
connection off to application code.