steps towards schema management for paymentserver db
This lets me add a new charge_id column to the vouchers table.
This lets us associate vouchers with transactions on our side.
This provides redundancy with the association on the stripe side.
With this redundancy we can let metadata mismatches go by
(maybe with logging/alerting)
instead of making them fail the whole process.