Fix the memory leak in getPublicKey
`newCString` and all of these libchallenge_bypass_ristretto functions require
that we do cleanup for them.  `withCAString` is the cleanup version of
`newCString` and `bracket` gives us a place to hang our other cleanup logic.

Also `CAString` is "C ASCII String" which is what we always want.  `CString`
is "C locale-aware String" :(